requirements.txt
django==4.2.6
djangorestframework
djangorestframework-simplejwt
drf-spectacular
psycopg2-binary
